requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
astroid==3.2.2
attrs==23.2.0
dill==0.3.8
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pylint==3.2.3
pytest==8.2.2
tomlkit==0.12.5
